Closes #7 Refactor modelflow core so it does not need numba to run.
Shrink requirements.txt to not include numba or graphviz to help avoid user errors.
