--- examples/mars/tests/integration_tests/test_integration.py
@@ -1,19 +1,0 @@
-# TODO
-# import sys
-# sys.path.insert(0, "../..")
-# from models import Location, Raddish, HabitatAtmosphere
-# from modelflow.modelflow import run_simulation
-
-
-# def test_basic_full_system():
-
-#     scenario = {
-#         "models": [
-#             Location(),
-#             Raddish(),
-#             HabitatAtmosphere(),
-#         ],
-#         "run_for_steps": 2,
-#         "time_per_step": "1hr" # TODO: Actually use this
-#     }
-#     run_simulation(scenario)
