+++ requirements.txt
@@ -1,18 +1,21 @@
+astroid==2.4.2
 attrs==20.2.0
 click==7.1.2
 Flask==1.1.2
 Flask-Cors==3.0.9
 iniconfig==1.1.1
+isort==5.6.4
 itsdangerous==1.1.0
 Jinja2==2.11.2
-llvmlite==0.34.0
+lazy-object-proxy==1.4.3
 MarkupSafe==1.1.1
-numba==0.51.2
+mccabe==0.6.1
 numpy==1.19.2
 packaging==20.4
 pandas==1.1.3
 pluggy==0.13.1
 py==1.9.0
+pylint==2.6.0
 pyparsing==2.4.7
 pytest==6.1.1
 python-dateutil==2.8.1
@@ -20,3 +23,4 @@ pytz==2020.1
 six==1.15.0
 toml==0.10.1
 Werkzeug==1.0.1
+wrapt==1.12.1
